deprecate buildSubgraphFromLabels!, getSubgraph, getSubgraphAroundNode, _copyIntoGraph!
